Fix Logout
